agent/pyproject.toml
[project]
name = "superior-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "anthropic>=0.45.2",
    "bs4>=0.0.2",
    "docker>=7.1.0",
    "duckduckgo-search>=7.3.0",
    "eth-typing>=5.1.0",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "langchain-community>=0.3.17",
    "langchain-openai>=0.3.5",
    "langchain>=0.3.18",
    "loguru>=0.7.3",
    "numpy>=2.2.3",
    "ollama>=0.4.7",
    "openai>=1.60.2",
    "peewee>=3.17.8",
    "pip>=25.0",
    "polars>=1.21.0",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "result>=0.17.0",
    "retry>=0.9.2",
    "tweepy>=4.15.0",
    "uvicorn>=0.34.0",
    "web3>=7.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
indent-width = 4

[tool.ruff.format]
indent-style = "tab"